Major Changes
I switched it to support erb views, and switched from a form to just a input text box, click or enter key work to send data. It communicates with the sql db, once you set it up. Just import the sql file and change the database name, username and password in the ruby file to match yours. Currently, it just gets the data from the browser and sends it to the db and also spits out the names entered.

We still need to:

"When the user enters a name your Javascript code must make an AJAX request (I strongly suggest you use jQuery) back to your Ruby (or Clojure) web app to get the data for that particular name." 

We still don't have an AJAX request. Maybe check out his code from earlier in the semester for a refresher.

"Your code must convert the 0's to 1001"

Shouldn't be too hard.

"The graphing must be done as shown, namely, with the most popular (rank 1) at the top."

"The Ruby web app will then look the name up in a SQL database and return the data or if not found an appropriate message."

We only have to figure out how to return the data to the graph or display an error message.

"You must sanitize both the input from Javascript web page, as well as, the data coming out of the database."

Then...
"Your HTML must validate as HTML5.
Any styling must be done as an external stylesheet that must validate as CSS3.
You must use Bootstrap (from CDN) to format your page."